mcscCTF25/misc/tg-bot/requirements.txt
python-dotenv
pyTelegramBotAPI
morse-audio-decoder
